extract unwanted functions from field_bgm
