+++ main.lsf
@@ -203,6 +203,7 @@ Static main
     Object src/script_pokemon_util.o
     Object src/terrain_attributes.o
     Object asm/unk_02054648.o
+    Object src/unk_02054E00.o
     Object src/field_bgm.o
     Object asm/unk_020551B8.o
     Object asm/unk_02055244.o
